Comment out default config os option
For some reason, after merging master into this branch, attempting to open a file caused a crash.
The OS section of the default config in app_config.go didn't get overridden by the user's config, and I'm not sure why.
I'm going to comment it out for now and see if the issue resurfaces in some other way later