fix panic when trying to open file at line by line panel
As described in #1187, when trying to open a file at the line by line
panel, a panic generated.

The issue appears to be that we were trying to open files with line
number append to it i.e. `file.go:123`, causing an error on the
execution of the command for opening the file.

I'm sure there's some context as to why it was implemented that way
so any feedback would be great :)

Fixes #1187