let arrow keys add lines to commit message window
Allows users to dynamically increase or decrease the height of the
commit message window by using the up & down arrows.  Pressing the down
arrow when the cursor is on the last line of the window increases the
height of the window.  Pressing the up arrow when the cursor is on the
last line, AND the last line is blank, decreases the height of the
window by one line.

Relates to: #28