Fix yellow/red coloring of pushed/unpushed commits in branch commits panel
The detection of the first pushed commit (commits above this are colored red,
below this yellow) didn't work for the sub-commits view of the branches panel.
The reason is that we were calling

  git merge-base refs/heads/mybranch refs/heads/mybranch@{u}

to detect the merge base; this doesn't work, the @{u} prefix only works for
short ref names, not full refs. So change the call to

  git merge-base refs/heads/mybranch mybranch@{u}