Extend test for moving patch into new commit
The patch to be moved to a new commit contains one file with modification, one
file that is added, and one file that is removed in this commit. This all works
already, but since one of these cases requires special handling when we
reimplement the "move to new commit" command later in this branch, it's good to
have complete test coverage for it.

Move the files into a directory to make it easier to create a custom patch for
all three at once with a single key press.