Pass --empty=keep to interactive rebase command
The --keep-empty flag only affects commits that were already empty when the
rebase starts. We also want to keep commits that become empty during the rebase;
one situation where this can happen is when moving a custom patch to an earlier
commit and there are no changes left in the source commit.