[gocui] Fix a bug in the code that clears to the end of the line
This code assumed that the length of the current line is the same as the current
x write position, which isn't necessarily true if we are overwriting existing
lines.

This hasn't been a problem so far because we would only use the erase-to-eol
feature in cases where the length of the lines either didn't change, or where we
called v.clear before. With this branch, we can run into it for example when
pressing "+" to expand the commits view and then going back.