allow verbose diff in rebase commit editor
There currently exists a "verbose" flag in the config that a user can
enable in order to show the diff when writing a commit message in the
editor of their choice. However, I discovered that this isn't the case
when you're rewording an existing commit through a rebase.

I find it's much easier to write meaningful commit messages during a
rebase if the diff is right there in front of you, so this commit adds
the ability for that verbose flag to affect the rebase commit editing as
well.

I thought about adding another flag to the config, but figured that was
a bit overkill since this is still to do with commit messages.

---

I've copied the `verboseFlag()` method from `CommitCommands` into
`LocalCommitsController` as there was no relation between the two (and
it's a very useful convenience method). This also means that it can be
used in `RebaseCommands` due to their relationship.